AudioSystem Event Queue
Turned the AudioSystem class into an Event Queue, using a ring buffer and a fixed size array to process sound requests. A service locator can use this audio system event queue as a service, so with a single notify call from anywhere in the codebase the right sound code will be executed. Next step would be to process all the audio request on a different thread.