Audio System Updates
Made it possible to add sounds in different ways, one where you specify the ID of the sound and one where you get an ID to call on that sound. Added extra functionality to fix the problem of playing the same sound multiple times in the same frame. There's checks to make sure the same sound request will only be processed after "x" amount of seconds.